# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Keuangan Simpadu API",
        "description": "Dokumentasi API Keuangan Simpadu",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/keringanan": {
            "get": {
                "tags": [
                    "Keringanan"
                ],
                "summary": "Tampilkan semua keringanan",
                "operationId": "a361a9683dbb6c246023149a3207c113",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id_keringanan": {
                                                "type": "integer"
                                            },
                                            "nim": {
                                                "type": "string"
                                            },
                                            "id_thn_ak": {
                                                "type": "string"
                                            },
                                            "jenis_keringanan": {
                                                "type": "string"
                                            },
                                            "jumlah_potongan": {
                                                "type": "integer"
                                            },
                                            "deskripsi_keringanan": {
                                                "type": "string"
                                            },
                                            "status_keringanan": {
                                                "type": "string"
                                            },
                                            "tgl_konfirmasi": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "id_tagihan": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Keringanan"
                ],
                "summary": "Tambah data keringanan",
                "operationId": "b2aa35b52f118ee83b11925d2f00b394",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nim",
                                    "id_thn_ak",
                                    "jenis_keringanan",
                                    "jumlah_potongan"
                                ],
                                "properties": {
                                    "nim": {
                                        "type": "string",
                                        "example": "1234567890123456"
                                    },
                                    "id_thn_ak": {
                                        "type": "string",
                                        "example": "20245"
                                    },
                                    "jenis_keringanan": {
                                        "type": "string",
                                        "example": "Potongan Prestasi"
                                    },
                                    "jumlah_potongan": {
                                        "type": "integer",
                                        "example": 500000
                                    },
                                    "deskripsi_keringanan": {
                                        "type": "string",
                                        "example": "Potongan karena juara olimpiade"
                                    },
                                    "status_keringanan": {
                                        "type": "string",
                                        "example": "Disetujui"
                                    },
                                    "tgl_konfirmasi": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-20"
                                    },
                                    "id_tagihan": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data keringanan berhasil ditambahkan"
                    }
                }
            }
        },
        "/api/keringanan/{id}": {
            "get": {
                "tags": [
                    "Keringanan"
                ],
                "summary": "Detail keringanan berdasarkan ID (beserta data mahasiswa dan nama tahun akademik dari microservice)",
                "operationId": "5040fa5468fadcd21a70b6112839938a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id_keringanan": {
                                            "type": "integer"
                                        },
                                        "nim": {
                                            "type": "string"
                                        },
                                        "mahasiswa": {
                                            "properties": {
                                                "nim": {
                                                    "type": "string"
                                                },
                                                "nama": {
                                                    "type": "string"
                                                },
                                                "prodi": {
                                                    "type": "string"
                                                },
                                                "jenjang": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "id_thn_ak": {
                                            "type": "string"
                                        },
                                        "nama_thn_ak": {
                                            "type": "string"
                                        },
                                        "jenis_keringanan": {
                                            "type": "string"
                                        },
                                        "jumlah_potongan": {
                                            "type": "integer"
                                        },
                                        "deskripsi_keringanan": {
                                            "type": "string"
                                        },
                                        "status_keringanan": {
                                            "type": "string"
                                        },
                                        "tgl_konfirmasi": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "id_tagihan": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Keringanan"
                ],
                "summary": "Update data keringanan",
                "operationId": "8007717214c098757fa7d76888387d31",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "jenis_keringanan": {
                                        "type": "string"
                                    },
                                    "jumlah_potongan": {
                                        "type": "integer"
                                    },
                                    "deskripsi_keringanan": {
                                        "type": "string"
                                    },
                                    "status_keringanan": {
                                        "type": "string"
                                    },
                                    "tgl_konfirmasi": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "id_tagihan": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data keringanan berhasil diupdate"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Keringanan"
                ],
                "summary": "Hapus data keringanan",
                "operationId": "0afdab058b3776b1e383405f410c0f41",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data keringanan berhasil dihapus"
                    }
                }
            }
        },
...
    },
...
}